refactor: renaming, simplify code
